add exceptionss
