+++ app/src/main/java/com/dt5gen/landmarkhelper/viewmodel/LandmarkViewModel.kt
@@ -1,5 +1,6 @@
 package com.dt5gen.landmarkhelper.viewmodel
 
+import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.dt5gen.landmarkhelper.model.Landmark
@@ -18,15 +19,34 @@ class LandmarkViewModel(private val apiService: ApiService) : ViewModel() {
     fun fetchLandmarks() {
         viewModelScope.launch {
             isLoading.value = true
+            errorMessage.value = null
             try {
                 val response = apiService.getLandmarks()
                 _landmarks.value = response
-                errorMessage.value = null
+                Log.d("LandmarkViewModel", "Fetched landmarks: $response") // Лог успешного получения данных
             } catch (e: Exception) {
-                errorMessage.value = e.message
+                errorMessage.value = "Failed to load landmarks"
+                Log.e("LandmarkViewModel", "Error fetching landmarks", e) // Лог ошибки
             } finally {
                 isLoading.value = false
             }
         }
     }
 }
+
+//fun fetchLandmarks() {
+//    viewModelScope.launch {
+//        _isLoading.value = true
+//        _errorMessage.value = null
+//        try {
+//            val response = apiService.getLandmarks()
+//            _landmarks.value = response
+//            Log.d("LandmarkViewModel", "Fetched landmarks: $response") // Лог успешного получения данных
+//        } catch (e: Exception) {
+//            _errorMessage.value = "Failed to load landmarks"
+//            Log.e("LandmarkViewModel", "Error fetching landmarks", e) // Лог ошибки
+//        } finally {
+//            _isLoading.value = false
+//        }
+//    }
+//}
